Converting to parcel from top-level html
this is going to help in a few ways. it can bundle the fonts, theoretically.
which is necessary for gdpr in europe.
And I can get hot reload out of it, which will be handy.

But it's being really hard.

The fake isn't working because stuff wasn't copied... I can move it into the fake.